# Website/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduAI Platform API",
        "description": "API Documentation for EduAI Platform",
        "contact": {
            "email": "support@eduai.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/api/v1/tools/animation": {
            "post": {
                "tags": [
                    "4. Tools"
                ],
                "summary": "Generate Video Animation",
                "operationId": "452841134acfa845f84d6a1cb55dbcfa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "prompt"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "integer"
                                    },
                                    "prompt": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Job Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job started"
                                        },
                                        "job_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/tools/audio": {
            "post": {
                "tags": [
                    "4. Tools"
                ],
                "summary": "Generate Audio Narration",
                "operationId": "78d7db3f00f459fec5a69d34247796aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "text"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "integer"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Job Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job started"
                                        },
                                        "job_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/tools/mindmap": {
            "post": {
                "tags": [
                    "4. Tools"
                ],
                "summary": "Generate Mind Map",
                "operationId": "a9f83587569f2e71df0b86117a00eb17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "topic"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "integer"
                                    },
                                    "topic": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Generation Complete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "job_id": {
                                            "type": "integer"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "succeeded"
                                        },
                                        "result": {
                                            "description": "Markdown content",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/tools/presentation": {
            "post": {
                "tags": [
                    "4. Tools"
                ],
                "summary": "Generate Presentation",
                "description": "Starts a background job to generate a presentation",
                "operationId": "c09c684588b941dc8f7e026de7737112",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "topic"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "topic": {
                                        "type": "string",
                                        "example": "The utility of AI in Education"
                                    },
                                    "style": {
                                        "type": "string",
                                        "example": "Modern"
                                    },
                                    "slide_count": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "instructions": {
                                        "type": "string",
                                        "example": "Include images"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Job Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job started"
                                        },
                                        "job_id": {
                                            "description": "Use this ID to check status",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Session Cookie Authentication",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}